+++ stac/CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - `Bbox` ([#303](https://github.com/stac-utils/stac-rs/pull/303))
 - Functions to create collections from items ([#304](https://github.com/stac-utils/stac-rs/pull/304))
 - Default implementation for `Version` ([#309](https://github.com/stac-utils/stac-rs/pull/309))
+- Experimental GeoParquet and GeoArrow support ([#316](https://github.com/stac-utils/stac-rs/pull/316))
 
 ### Changed